Add `{StorageDb, Store}::teardown()`. Closes #3014.
`teardown` explicitly closes the underlying database connection.
This is unnecessary to call from Rust, where dropping a `StorageDb`
automatically closes the connection, but is helpful for Desktop, where
the store is accessed from a background thread and teardown is
asynchronous.

Unlike dropping a `StorageDb`, which panics if the database can't be
closed, `teardown` leaks the connection. This is important for Desktop,
because we don't want to crash the process if closing the connection
fails. Since we only expect to call `teardown` at shutdown, this is
fine, as the resources will be shortly reclaimed by the OS.